pyproject.toml
[project]
name = "llamaindex-study"
version = "0.2.0"
description = "LlamaIndex 学习项目 - 现代化 RAG 应用开发"
readme = "README.md"
authors = [
    {name = "lpchg1992", email = "huitailang_luo@126.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    # LlamaIndex 核心包 - 升级到 v0.14.x
    "llama-index-core>=0.14.0",
    "llama-index-llms-openai>=0.3.0",
    "llama-index-embeddings-ollama>=0.3.0",
    "llama-index-llms-ollama>=0.10.0",
    # 向量存储包
    "llama-index-vector-stores-lancedb>=0.5.0,<0.6.0",
    "llama-index-vector-stores-chroma>=0.5.0,<0.6.0",
    # qdrant 暂时禁用 - Python 3.14 兼容性尚未支持
    # "llama-index-vector-stores-qdrant>=0.9.0,<0.11.0",
    # LanceDB - 更新到兼容版本
    "lancedb>=0.6.0,<1.0.0",
    # 文档解析和读取
    "pypdf>=6.9.2,<7.0.0",
    "llama-index-readers-file>=0.6.0,<0.7.0",
    "reportlab>=4.4.10,<5.0.0",
    "python-docx>=1.2.0,<2.0.0",
    "openpyxl>=3.1.5,<4.0.0",
    "python-pptx>=1.0.2,<2.0.0",
    "docx2txt>=0.9,<0.10",
    # 其他依赖
    "chromadb>=1.2.1,<2.0.0",
    # qdrant 暂时禁用
    # "qdrant-client>=1.15.1,<2.0.0",
    "scikit-learn>=1.8.0,<2.0.0",
    "fastapi>=0.135.2,<0.136.0",
    "markdown>=3.7",
    "uvicorn>=0.42.0,<0.43.0",
    "httpx>=0.28.1,<0.29.0",
    "sqlalchemy>=2.0.38,<3.0.0",
    "python-dotenv>=1.0.0",
    # Semantic chunking
    "llama-index-packs-node-parser-semantic-chunking>=0.5.1,<0.6.0",
    # Hybrid search
    "llama-index-retrievers-bm25>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
llamaindex-study = "kb_cli.__main__:main"
